Use the builtin rvm script instead of the shell function to install
Using the shell function makes rvm assume it's an interactive install.
The rvm requirements page shows up in 'less' which is unconfigurable.
This is a workaround to avoid veewee hanging on a less process that you
can't quit out of without SSHing into the box and killing the less
process itself.